+++ flowmachine/tests/test_spatial_aggregate.py
@@ -3,7 +3,11 @@
 # file, You can obtain one at http://mozilla.org/MPL/2.0/.
 
 from flowmachine.core import make_spatial_unit
-from flowmachine.features import ModalLocation, daily_location, SubscriberPhoneType
+from flowmachine.features import (
+    ModalLocation,
+    daily_location,
+    SubscriberHandsetCharacteristic,
+)
 from flowmachine.features.utilities.spatial_aggregates import JoinedSpatialAggregate
 from flowmachine.features.subscriber.daily_location import locate_subscribers
 from flowmachine.utils import list_of_dates
@@ -49,8 +53,10 @@ def test_can_be_aggregated_admin3_distribution(get_dataframe):
         spatial_unit=make_spatial_unit("admin", level=3),
         method="most-common",
     )
-    metric = SubscriberPhoneType("2016-01-01", "2016-01-02")
+    metric = SubscriberHandsetCharacteristic(
+        "2016-01-01", "2016-01-02", characteristic="hnd_type"
+    )
     agg = JoinedSpatialAggregate(metric=metric, locations=locations, method="dist")
     df = get_dataframe(agg)
     assert ["pcod", "metric", "key", "value"] == list(df.columns)
-    assert df[df.metric == "handset_type"].groupby("pcod").sum().value.unique() == 1
+    assert df[df.metric == "value"].groupby("pcod").sum().value.unique() == 1