+++ flowmachine/tests/test_subscriber_tacs.py
@@ -14,19 +14,11 @@ from flowmachine.features.subscriber import (
     SubscriberHandset,
     SubscriberHandsets,
     SubscriberHandsetCharacteristic,
-    SubscriberPhoneType,
 )
 
 
 @pytest.mark.parametrize(
-    "query",
-    [
-        SubscriberTACs,
-        SubscriberTAC,
-        SubscriberHandset,
-        SubscriberHandsets,
-        SubscriberPhoneType,
-    ],
+    "query", [SubscriberTACs, SubscriberTAC, SubscriberHandset, SubscriberHandsets]
 )
 def test_column_names(query, get_dataframe):
     """Test that column_names attribute matches columns from calling head"""
@@ -105,25 +97,6 @@ def test_subscriber_handset(get_dataframe):
         "subscriber"
     )
     assert tc.loc["1p4MYbA1Y4bZzBQa"].model == "LB-01"
-
-
-def test_subscriber_phonetype(get_dataframe):
-    """Check that correct smart/feature label is returned."""
-    assert (
-        get_dataframe(SubscriberPhoneType("2016-01-01", "2016-01-07"))
-        .set_index("subscriber")
-        .loc["038OVABN11Ak4W5P"]
-        .handset_type
-        == "Smart"
-    )
-
-    assert (
-        get_dataframe(SubscriberPhoneType("2016-01-01", "2016-01-07", method="last"))
-        .set_index("subscriber")
-        .loc["YMBqRkzbbxGkX3zA"]
-        .handset_type
-        == "Feature"
-    )
 
 
 def test_subscriber_handset_characteristic(get_dataframe):